tryon_backend/requirements.txt
gradio_client==1.4.2
fastapi
fastapi[standard]
uvicorn
python-dotenv
ngrok
pyngrok
opencv-python
opencv-contrib-python
twilio
requests
cloudinary
